calc memoryload as a percentage now

+++ tools/xen_ressource_manager/rm_xenstore.h
@@ -5,7 +5,7 @@ int RM_XENSTORE_init(void);
 
 void RM_XENSTORE_close(void);
 
-int RM_XENSTORE_read_domain_memload(int domid);
+double RM_XENSTORE_read_domain_memload(int domid);
 
 double RM_XENSTORE_read_domain_cpuload(int domid);
